refactor: move server/config tests into test/server/config
