+++ test/server/config/header-title.spec.ts
@@ -10,9 +10,9 @@ import {
   TITLE_REGEN_EVERY_TURNS,
   VIEW_TITLE_REGEN_TURNS,
   MAX_TITLE_CHARS,
-} from "./header-title.js";
-import type { RunClaude } from "../session/command-summary.js";
-import type { ConversationTurn } from "../session/transcript.js";
+} from "../../../../server/../../server/config/header-title.js";
+import type { RunClaude } from "../../../../server/session/command-summary.js";
+import type { ConversationTurn } from "../../../../server/session/transcript.js";
 
 const line = (o: unknown) => JSON.stringify(o);
 const ok = (stdout: string): RunClaude => vi.fn(async () => ({ stdout, stderr: "", code: 0 }));
